+++ examples/apache-storm/src/test/java/nl/basjes/parse/httpdlog/storm/ParserBoltTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.basjes.storm.bolt;
+package nl.basjes.parse.httpdlog.storm;
 
 import org.apache.storm.LocalCluster;
 import org.apache.storm.generated.StormTopology;
@@ -38,9 +38,11 @@ import java.io.Serializable;
 import java.util.HashMap;
 import java.util.Map;
 
-@SuppressWarnings("serial")
 public class ParserBoltTest implements Serializable {
     // ========================================================================
+
+    private static final String INPUT_FIELD_NAME = "apacheLogline";
+    private static final String OUTPUT_FIELD_NAME = "resultingRecord";
 
     private static final Logger LOG = LoggerFactory.getLogger(ParserBoltTest.class);
 
@@ -52,21 +54,12 @@ public class ParserBoltTest implements Serializable {
         }
 
         public void nextTuple() {
-            String logline = "84.105.31.162 - - [05/Sep/2010:11:27:50 +0200] "
-                    + "\"GET /fotos/index.html?img=geboorte-kaartje&foo=foofoo&bar=barbar HTTP/1.1\" 200 23617 "
-                    + "\"http://www.google.nl/imgres?imgurl=http://daniel_en_sander.basjes.nl/fotos/geboorte-kaartje/"
-                    + "geboortekaartje-binnenkant.jpg&imgrefurl=http://daniel_en_sander.basjes.nl/fotos/geboorte-kaartje"
-                    + "&usg=__LDxRMkacRs6yLluLcIrwoFsXY6o=&h=521&w=1024&sz=41&hl=nl&start=13&zoom=1&um=1&itbs=1&"
-                    + "tbnid=Sqml3uGbjoyBYM:&tbnh=76&tbnw=150&prev=/images%3Fq%3Dbinnenkant%2Bgeboortekaartje%26um%3D1%26hl%3D"
-                    + "nl%26sa%3DN%26biw%3D1882%26bih%3D1014%26tbs%3Disch:1\" "
-                    + "\"Mozilla/5.0 (Macintosh; U; Intel Mac OS X 10_6_4; nl-nl) AppleWebKit/533.17.8 (KHTML, like Gecko) "
-                    + "Version/5.0.1 Safari/533.17.8\"";
-            collector.emit(new Values(logline));
+            collector.emit(new Values(TestCase.getInputLine()));
             Utils.sleep(1000L);
         }
 
         public void declareOutputFields(OutputFieldsDeclarer declarer) {
-            declarer.declare(new Fields("apachelogline"));
+            declarer.declare(new Fields(INPUT_FIELD_NAME));
         }
     }
 
@@ -76,16 +69,9 @@ public class ParserBoltTest implements Serializable {
         @Override
         public void execute(Tuple tuple, BasicOutputCollector collector) {
             Fields fields = tuple.getFields();
-            Assert.assertEquals(7, fields.size());
-            Assert.assertEquals("2010", tuple.getStringByField("year"));
-            Assert.assertEquals("9", tuple.getStringByField("month"));
-            Assert.assertEquals("5", tuple.getStringByField("day"));
-            Assert.assertEquals("11", tuple.getStringByField("hour"));
-            Assert.assertEquals("27", tuple.getStringByField("minute"));
-            Assert.assertEquals("50", tuple.getStringByField("second"));
-            Assert.assertEquals("Mozilla/5.0 (Macintosh; U; Intel Mac OS X 10_6_4; nl-nl) " +
-                                "AppleWebKit/533.17.8 (KHTML, like Gecko) Version/5.0.1 " +
-                                "Safari/533.17.8", tuple.getStringByField("useragent"));
+            Assert.assertEquals(1, fields.size());
+            TestRecord record = (TestRecord) tuple.getValueByField(OUTPUT_FIELD_NAME);
+            record.assertIsValid();
             LOG.info("Test passed");
         }
 
@@ -97,21 +83,13 @@ public class ParserBoltTest implements Serializable {
 
     // ========================================================================
     @Test
-    public void runRest() throws InterruptedException {
+    public void runRest() throws InterruptedException, NoSuchMethodException {
         TopologyBuilder builder = new TopologyBuilder();
 
         // ----------
         builder.setSpout("Spout", new TestApacheLogsSpout());
         // ----------
-        String logformat = "%h %l %u %t \"%r\" %>s %b \"%{Referer}i\" \"%{User-Agent}i\"";
-        ApacheHttpdLoglineParserBolt parserBolt = new ApacheHttpdLoglineParserBolt(logformat, "apachelogline");
-        parserBolt.requestField("TIME.YEAR:request.receive.time.year",          "year");
-        parserBolt.requestField("TIME.MONTH:request.receive.time.month",        "month");
-        parserBolt.requestField("TIME.DAY:request.receive.time.day",            "day");
-        parserBolt.requestField("TIME.HOUR:request.receive.time.hour",          "hour");
-        parserBolt.requestField("TIME.MINUTE:request.receive.time.minute",      "minute");
-        parserBolt.requestField("TIME.SECOND:request.receive.time.second",      "second");
-        parserBolt.requestField("HTTP.USERAGENT:request.user-agent",            "useragent");
+        HttpdLoglineParserBolt parserBolt = new HttpdLoglineParserBolt(TestCase.getLogFormat(), INPUT_FIELD_NAME, OUTPUT_FIELD_NAME);
 
         builder.setBolt("Parser", parserBolt, 1).shuffleGrouping("Spout");
         // ----------